image/requirements.txt
tensorflow
numpy
opencv-python
fastapi
uvicorn
Pillow
python-multipart